add futures_contract_info_czce
