+++ docs/data/futures/futures.md
@@ -2325,7 +2325,7 @@ print(futures_contract_info_shfe_df)
 [276 rows x 8 columns]
 ```
 
-#### 郑州商品交易所
+#### 大连商品交易所
 
 接口: futures_contract_info_dce
 
@@ -2378,6 +2378,217 @@ print(futures_contract_info_dce_df)
 218  豆油  y2412    10     2.0  2023-12-15  2024-12-13  2024-12-18
 219  豆油  y2501    10     2.0  2024-01-16  2025-01-15  2025-01-20
 [220 rows x 7 columns]
+```
+
+#### 郑州商品交易所
+
+接口: futures_contract_info_czce
+
+目标地址: http://www.czce.com.cn/cn/jysj/cksj/H770322index_1.htm
+
+描述: 郑州商品交易所-交易数据-参考数据
+
+限量: 单次返回指定 date 的期货合约信息数据
+
+输入参数
+
+| 名称   | 类型  | 描述                   |
+|------|-----|----------------------|
+| date | str | date="20240228"; 交易日 |
+
+输出参数
+
+| 名称                        | 类型      | 描述 |
+|---------------------------|---------|----|
+| 产品名称                      | object  | -  |
+| 合约代码                      | object  | -  |
+| 产品代码                      | object  | -  |
+| 产品类型                      | object  | -  |
+| 交易所MIC编码                  | object  | -  |
+| 交易场所                      | object  | -  |
+| 交易时间节假日除外                 | object  | -  |
+| 交易国家ISO编码                 | object  | -  |
+| 交易币种ISO编码                 | object  | -  |
+| 结算币种ISO编码                 | object  | -  |
+| 到期时间待国家公布2025年节假日安排后进行调整  | object  | -  |
+| 结算方式                      | object  | -  |
+| 挂牌频率                      | object  | -  |
+| 最小变动价位                    | object  | -  |
+| 最小变动价值                    | object  | -  |
+| 交易单位                      | object  | -  |
+| 计量单位                      | object  | -  |
+| 最大下单量                     | object  | -  |
+| 日持仓限额期货公司会员不限仓            | object  | -  |
+| 大宗交易最小规模                  | object  | -  |
+| 是否受CESR监管                 | object  | -  |
+| 是否为灵活合约                   | object  | -  |
+| 上市周期该产品的所有合约月份            | object  | -  |
+| 交割通知日                     | object  | -  |
+| 第一交易日                     | object  | -  |
+| 最后交易日待国家公布2025年节假日安排后进行调整 | object  | -  |
+| 交割结算日                     | object  | -  |
+| 月份代码                      | object  | -  |
+| 年份代码                      | object  | -  |
+| 最后交割日                     | object  | -  |
+| 车（船）板最后交割日                | object  | -  |
+| 合约交割月份本合约交割月份             | object  | -  |
+| 交易保证金率                    | object  | -  |
+| 涨跌停板                      | object  | -  |
+| 费用币种ISO编码                 | object  | -  |
+| 交易手续费                     | float64 | -  |
+| 手续费收取方式                   | object  | -  |
+| 交割手续费                     | float64 | -  |
+| 平今仓手续费                    | float64 | -  |
+| 交易限额                      | float64 | -  |
+
+接口示例
+
+```python
+import akshare as ak
+
+futures_contract_info_czce_df = ak.futures_contract_info_czce(date="20240228")
+print(futures_contract_info_czce_df)
+```
+
+数据示例
+
+```
+      产品名称  合约代码 产品代码 产品类型 交易所MIC编码  ...  交易手续费 手续费收取方式 交割手续费 平今仓手续费  交易限额
+0    鲜苹果期货  AP403   AP   期货     XZCE  ...    5.0     绝对值   0.0   20.0   NaN
+1    鲜苹果期货  AP404   AP   期货     XZCE  ...    5.0     绝对值   0.0   20.0   NaN
+2    鲜苹果期货  AP405   AP   期货     XZCE  ...    5.0     绝对值   0.0   20.0   NaN
+3    鲜苹果期货  AP410   AP   期货     XZCE  ...    5.0     绝对值   0.0   20.0   NaN
+4    鲜苹果期货  AP411   AP   期货     XZCE  ...    5.0     绝对值   0.0   20.0   NaN
+..     ...    ...  ...  ...      ...  ...    ...     ...   ...    ...   ...
+213  动力煤期货  ZC410   ZC   期货     XZCE  ...  150.0     绝对值   0.0  150.0  20.0
+214  动力煤期货  ZC411   ZC   期货     XZCE  ...  150.0     绝对值   0.0  150.0  20.0
+215  动力煤期货  ZC412   ZC   期货     XZCE  ...  150.0     绝对值   0.0  150.0  20.0
+216  动力煤期货  ZC501   ZC   期货     XZCE  ...  150.0     绝对值   0.0  150.0  20.0
+217  动力煤期货  ZC502   ZC   期货     XZCE  ...  150.0     绝对值   0.0  150.0  20.0
+[218 rows x 40 columns]
+```
+
+#### 广州期货交易所
+
+接口: futures_contract_info_gfex
+
+目标地址: http://www.gfex.com.cn/gfex/hyxx/ywcs.shtml
+
+描述: 广州期货交易所-业务/服务-合约信息
+
+限量: 单次返回最近交易日的期货合约信息数据
+
+输入参数
+
+| 名称 | 类型 | 描述 |
+|----|----|----|
+| -  | -  | -  |
+
+输出参数
+
+| 名称     | 类型     | 描述 |
+|--------|--------|----|
+| 品种     | object | -  |
+| 合约代码   | object | -  |
+| 交易单位   | int64  | -  |
+| 最小变动单位 | int64  | -  |
+| 开始交易日  | object | -  |
+| 最后交易日  | object | -  |
+| 最后交割日  | object | -  |
+
+接口示例
+
+```python
+import akshare as ak
+
+futures_contract_info_gfex_df = ak.futures_contract_info_gfex()
+print(futures_contract_info_gfex_df)
+```
+
+数据示例
+
+```
+     品种    合约代码  交易单位  最小变动单位       开始交易日       最后交易日       最后交割日
+0   碳酸锂  lc2403     1      50  2023-07-21  2024-03-14  2024-03-19
+1   碳酸锂  lc2404     1      50  2023-07-21  2024-04-16  2024-04-19
+2   碳酸锂  lc2405     1      50  2023-07-21  2024-05-17  2024-05-22
+3   碳酸锂  lc2406     1      50  2023-07-21  2024-06-17  2024-06-20
+4   碳酸锂  lc2407     1      50  2023-07-21  2024-07-12  2024-07-17
+5   碳酸锂  lc2408     1      50  2023-08-15  2024-08-14  2024-08-19
+6   碳酸锂  lc2409     1      50  2023-09-15  2024-09-13  2024-09-20
+7   碳酸锂  lc2410     1      50  2023-10-23  2024-10-21  2024-10-24
+8   碳酸锂  lc2411     1      50  2023-11-15  2024-11-14  2024-11-19
+9   碳酸锂  lc2412     1      50  2023-12-15  2024-12-13  2024-12-18
+10  碳酸锂  lc2501     1      50  2024-01-16  2025-01-15  2025-01-20
+11  碳酸锂  lc2502     1      50  2024-02-23  2025-02-24  2025-02-27
+12  工业硅  si2403     5       5  2023-03-15  2024-03-14  2024-03-19
+13  工业硅  si2404     5       5  2023-08-15  2024-04-16  2024-04-19
+14  工业硅  si2405     5       5  2023-09-15  2024-05-17  2024-05-22
+15  工业硅  si2406     5       5  2023-10-23  2024-06-17  2024-06-20
+16  工业硅  si2407     5       5  2023-11-15  2024-07-12  2024-07-17
+17  工业硅  si2408     5       5  2023-12-15  2024-08-14  2024-08-19
+18  工业硅  si2409     5       5  2024-01-16  2024-09-13  2024-09-20
+19  工业硅  si2410     5       5  2024-02-23  2024-10-21  2024-10-24
+```
+
+#### 中国金融期货交易所
+
+接口: futures_contract_info_cffex
+
+目标地址: http://www.gfex.com.cn/gfex/hyxx/ywcs.shtml
+
+描述: 中国金融期货交易所-数据-交易参数
+
+限量: 单次返回指定 date 的期货合约信息数据
+
+输入参数
+
+| 名称   | 类型  | 描述                   |
+|------|-----|----------------------|
+| date | str | date="20240228"; 交易日 |
+
+输出参数
+
+| 名称    | 类型      | 描述 |
+|-------|---------|----|
+| 合约代码  | object  | -  |
+| 合约月份  | object  | -  |
+| 挂盘基准价 | float64 | -  |
+| 上市日   | object  | -  |
+| 最后交易日 | object  | -  |
+| 涨停板幅度 | object  | -  |
+| 跌停板幅度 | object  | -  |
+| 涨停板价位 | float64 | -  |
+| 跌停板价位 | float64 | -  |
+| 持仓限额  | int64   | -  |
+| 品种    | object  | -  |
+| 查询交易日 | object  | -  |
+
+接口示例
+
+```python
+import akshare as ak
+
+futures_contract_info_cffex_df = ak.futures_contract_info_cffex(date="20240228")
+print(futures_contract_info_cffex_df)
+```
+
+数据示例
+
+```
+        合约代码    合约月份    挂盘基准价 上市日  ...  跌停板价位 持仓限额 品种  查询交易日
+0           TS2409  2409  100.974  2023-12-11  ...  101.056  2000  TS  2024-02-28
+1           TS2406  2406  101.065  2023-09-11  ...  101.020  2000  TS  2024-02-28
+2           TS2403  2403  101.145  2023-06-12  ...  100.858  2000  TS  2024-02-28
+3           TL2409  2409   99.330  2023-12-11  ...  102.990  2000  TL  2024-02-28
+4           TL2406  2406   98.750  2023-09-11  ...  102.870  2000  TL  2024-02-28
+..             ...   ...      ...         ...  ...      ...   ...  ..         ...
+831  HO2403-C-2075  2403  193.200  2023-12-18  ...   83.800  1200  HO  2024-02-28
+832  HO2403-C-2050  2403  253.400  2023-12-06  ...  115.000  1200  HO  2024-02-28
+833  HO2403-C-2025  2403  234.800  2023-12-18  ...  133.600  1200  HO  2024-02-28
+834  HO2403-C-2000  2403  286.200  2023-12-08  ...  158.600  1200  HO  2024-02-28
+835  HO2403-C-1975  2403  241.200  2024-01-18  ...  189.600  1200  HO  2024-02-28
+[836 rows x 12 columns]
 ```
 
 ### 期货行情数据
